My git goofed so I had to delete everything and redo it..
removed a lot of the wasteland med spawners and switched to static spawned items. Removed most of the seeds in the farm, added advanced workbench at the very end of khans bunker. Justification for this is, Khans "down river" is harder than both ncr and legions combined and past f13 servers khans had it in their down river. I also made sure the walls around said advanced workbench were indestructible so it couldnt be cheesed, forcing players to run the whole bunker.