Renegades are back baby!
Added back the big renegades bunker. Made it harder and gave it better loot. The loot can be adjusted later when we see how easy or hard it is for factions and players to do. Also when the loot pools are more balanced. I also fixed a few little bugs from my past mapping pr's.